frontend and backend connection work and added playlist page

+++ backend/controllers/playlist.controller.js
@@ -1,25 +1,26 @@
 const { findByIdAndDelete } = require("../models/artist.model");
 const Playlist = require("../models/playlist.model");
 const USER = require("../models/user.model");
+const { trace } = require("../routes/playlist.routes");
 
 
 async function getPlaylist(req, res, next) {
     try {
-        const { search, limit = 10, page = 1, all ,top} = req.query
+        const { search, limit = 10, page = 1, all, top } = req.query
         const query = search ? { title: RegExp(search, 'i') } : {};
 
         if (all == "true") {
-            const playlist = await Playlist.find({})
-            return res.status(200).json({ message: "All Playlist", playlist })
+            const playlists = await Playlist.find({ isPrivate: false }).populate('songs')
+            return res.status(200).json({ message: "All Playlist", playlists })
         }
-        if(top == 'true'){
-            const playlists = await Playlist.find().sort({createdAt:-1}).limit(Number(limit))
+        if (top == 'true') {
+            const playlists = await Playlist.find({isPrivate: false}).sort({ createdAt: -1 }).limit(Number(limit))
             return res.status(200).json({ message: "Top Playlist", playlists })
         }
-        const playlists = await Playlist.find(query).skip((page - 1) * limit).limit(Number(limit))
+        const playlists = await Playlist.find({ ...query, isPrivate: false }).skip((page - 1) * limit).limit(Number(limit))
 
-        if (!playlists) {
-            return res.status(404).json({ message: "Failed to Get songs" })
+        if(!playlists){
+            return res.status(404).json({ message: "Playlist Not Found" })
         }
 
         res.status(200).json({ message: `Page ${page} playlist`, playlists })
@@ -29,16 +30,70 @@ async function getPlaylist(req, res, next) {
     }
 }
 
+async function getPlaylisById(req,res,next) {
+    try{
+        const id = req.params.playlistId;
+        console.log("id is :",id)
+        const playlist = await Playlist.findById(id).populate('songs');
+        if(!playlist){
+            return res.status(404).json({ message: "Playlist Not Found" })
+        }
+        res.status(200).json({message:`Playlist by id ${id}`,playlist})
+
+    }catch(error){
+        console.error("Error in Get Playlist By Id handeler : ", error.message)
+        res.status(500).json({ message: "Internal Server Error", error: error.message })
+    }
+}
+
+async function getUserPlaylist(req, res, next) {
+    try {
+        const { saved, yours } = req.query;
+        let userCreatedPlaylist = [];
+        let userSavedPlaylist = [];
+
+        // Ensure user exists
+        const user = await USER.findById(req.userId);
+        if (!user) {
+            return res.status(404).json({ message: "User not found" });
+        }
+
+        if (saved == 'true') {
+            const populatedPlaylist = await user.populate('playlist').then(u => u.playlist)
+            userSavedPlaylist = populatedPlaylist.filter(playlist => String(playlist.author) !== String(req.userId))
+        }
+
+        if (yours == 'true') {
+            userCreatedPlaylist = (await Playlist.find({ author: req.userId }))
+        }
+
+        res.status(200).json({ message: "User Playlists", userCreatedPlaylist, userSavedPlaylist })
+    } catch (error) {
+        console.error("Error in Get User Playlist handeler : ", error.message)
+        res.status(500).json({ message: "Internal Server Error", error: error.message })
+    }
+}
+
 async function createPlaylist(req, res, next) {
     try {
         const userId = req.userId;
         const { name, image, isPrivate } = req.body;
+
+        // Ensure user exists
+        const user = await USER.findById(userId);
+        if (!user) {
+            return res.status(404).json({ message: "User not found" });
+        }
+
         const playlist = await Playlist.create({
             author: userId,
             name,
             image,
             isPrivate
         })
+
+        user.playlist.push(playlist._id)
+        await user.save()
 
         res.status(201).json({ message: "Playlist Created Sucessfully", playlist })
     } catch (error) {
@@ -51,10 +106,20 @@ async function createPlaylist(req, res, next) {
 async function deletePlaylist(req, res, next) {
     try {
         const playlistId = req.params.playlistId;
+        // Ensure user exists
+        const user = await USER.findById(userId);
+        if (!user) {
+            return res.status(404).json({ message: "User not found" });
+        }
         const playlist = await findByIdAndDelete(playlistId)
         if (!playlist) {
             return res.status(404).json({ message: "Playlist Not Found" })
         }
+
+        let playlistIndex = user.playlist.findIndex(playlist._id)
+        user.playlist.slice(playlistIndex, 1);
+        await user.save()
+
         res.status(200).json({ message: "Playlist Deleted Sucessfully!", playlist })
     } catch (error) {
         console.error("Error in Delete Playlist handeler : ", error.message)
@@ -156,5 +221,7 @@ module.exports = {
     updatePlaylist,
     getPlaylist,
     toggleSongInPlaylist,
-    togglePlaylistSave
+    togglePlaylistSave,
+    getUserPlaylist,
+    getPlaylisById
 }
